git-svn-id: http://svn.seqan.de/seqan/trunk@161 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/system/file_sync.h
@@ -280,7 +280,7 @@ namespace SEQAN_NAMESPACE_MAIN
 		inline FilePtr seek(FilePtr pos, int origin = SEEK_SET) const {
             FilePtr result = ::lseek(handle, pos, origin);
 //			#ifdef SEQAN_DEBUG
-				if (result < 0) printf("seek returned %d %s\n", result, strerror(errno));
+				if (result < 0) printf("seek returned %d %s\n", (int)result, strerror(errno));
 //			#endif
 			return result;
 		}
